LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/opaque-response-preloaded.https-expected.txt


FAIL Opaque responses should not be reused for XHRs, loading case assert_equals: expected "PASS" but got "FAIL: {\"error\": {\"message\": \"\", \"code\": 404}}"
FAIL Opaque responses should not be reused for XHRs, done case assert_equals: expected "PASS" but got "FAIL: {\"error\": {\"message\": \"\", \"code\": 404}}"
